requirements.txt
paho-mqtt==1.6.1
pyserial==3.5
pyyaml==6.0.1
streamlit==1.33.0
plotly==5.20.0
pandas==2.1.4
pytest==7.4.4
